backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.6
defusedxml==0.8.0rc2
Django==5.0.4
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-filter==24.2
django-jalali==6.2.1
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.0.0
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
psycopg==3.1.18
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
